+++ fourtoon-cookie/src/apis/api.ts
@@ -1,4 +1,4 @@
-import { JwtError } from "../error/JwtError";
+import { JwtError } from "../types/error/JwtError";
 import { JWTToken } from "../types/jwt";
 import { API_METHOD_TYPE, API_STATUS } from "../constants/api";
 import { useJwtStore } from "../hooks/store/jwt";
