+++ fourtoon-cookie/src/hooks/server/member.ts
@@ -1,7 +1,7 @@
 import { useMutation, useQuery, useQueryClient } from "react-query";
 import { deleteMember, getMember, postMember } from "../../apis/member"
 import { Member } from "../../types/member";
-import { JwtError } from "../../error/JwtError";
+import { JwtError } from "../../types/error/JwtError";
 import { useJwtStore } from "../store/jwt";
 
 export const useMember = () => {
