+++ fourtoon-cookie/src/types/error/JwtError.ts
